requirements.txt
langchain-community>=0.0.1
langchain-openai>=0.0.1
langchain-core>=0.0.1
PyMuPDF>=1.20.0
